pyproject.toml
[project]
name = "dbt-with-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "acryl-datahub-airflow-plugin>=1.3.1.7",
    "acryl-datahub[datahub-rest,dbt]>=1.3.1.7",
    "apache-airflow==2.10.5",
    "apache-airflow-providers-docker>=4.4.5",
    "apache-airflow-providers-postgres>=6.4.1",
    "dbt-core>=1.11.2",
    "dbt-postgres==1.9.1",
    "duckdb>=1.4.3",
    "google-auth>=2.47.0",
    "gspread-pandas>=3.3.0",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
]
